refactor_app_structure: move parquet read and test filters
